% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_font_cache_dir.R
\name{clear_font_cache_dir}
\alias{clear_font_cache_dir}
\title{Clear Font Cache}
\usage{
clear_font_cache_dir(confirm = TRUE)
}
\arguments{
\item{confirm}{(\code{logical(1)}) Whether to ask for confirmation before deleting cached files. (default: TRUE)}
}
\value{
(\code{logical(1)}) TRUE if cache was cleared successfully, FALSE otherwise.
}
\description{
Removes cached Font Awesome fonts to free up space or force re-download.
}
