% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_download.R
\name{fa_download_single}
\alias{fa_download_single}
\title{Download Single Font Awesome Font Type}
\usage{
fa_download_single(font_type, version, force_download, quiet)
}
\arguments{
\item{font_type}{(\code{character(1)}) Font type to download: "brands" or "free_solid"}

\item{version}{(\code{character(1)}) Font Awesome version to download in semantic versioning format}

\item{force_download}{(\code{logical(1)}) Whether to force re-download even if cached}

\item{quiet}{(\code{logical(1)}) Whether to suppress download messages}
}
\value{
(\code{character(1) | NULL}) Character string with path to the downloaded font file, or NULL if download failed.
}
\description{
Helper function to download a specific Font Awesome font type.
}
\keyword{internal}
