% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_font_cache_dir.R
\name{get_font_cache_dir}
\alias{get_font_cache_dir}
\title{Get Font Cache Directory}
\usage{
get_font_cache_dir()
}
\value{
(\code{character(1)}) Absolute path to the Font Awesome fonts cache directory.
}
\description{
Returns the package directory where Font Awesome fonts are/should be cached.
}
\examples{
\donttest{
# Get font cache directory
cache_dir <- get_font_cache_dir()
print(cache_dir)
}

}
